Moved API to indipendent module
