+++ src/api/index.js
@@ -1,0 +1,3 @@
+import { getAlbum } from './api';
+
+export { getAlbum };
